+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/config/inherit_configuration.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Symfony\Component\DependencyInjection\Loader\Configurator;
+
+return function (ContainerConfigurator $container) {
+    $services = $container->services();
+
+    $services->defaults()->inheritConfiguration(false);
+
+    $services->set('foo', 'Foo');
+    $services->set('bar', 'Bar')->inheritConfiguration();
+};